refac delete template & delete receiver into separate actions
